Rename split operation and package to swap
