+++ lib/src/main/kotlin/me/tb/cashuclient/swap/SwapRequest.kt
@@ -3,14 +3,14 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the ./LICENSE file.
  */
 
-package me.tb.cashuclient.split
+package me.tb.cashuclient.swap
 
 import kotlinx.serialization.Serializable
 import me.tb.cashuclient.types.BlindedMessage
 import me.tb.cashuclient.types.Proof
 
 @Serializable
-public data class SplitRequest(
+public data class SwapRequest(
     public val proofs: Proof,
     public val outputs: List<BlindedMessage>
 )
